+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - UI for rendering dashboard with email tracking statistics
 - password protection for UI
+- endpoints for resending and deleting messages
 
 ## [0.2.3] - 2022-09-16
 ### Modified